move analyzer internals to analysis package
fixes #61
