+++ sudachi/src/plugin/path_rewrite/join_numeric/mod.rs
@@ -18,11 +18,11 @@ use serde::Deserialize;
 use serde_json::Value;
 
 use self::numeric_parser::NumericParser;
+use crate::analysis::{lattice::Lattice, node::Node};
 use crate::config::Config;
 use crate::dic::category_type::CategoryType;
 use crate::dic::grammar::Grammar;
 use crate::input_text::Utf8InputText;
-use crate::lattice::{node::Node, Lattice};
 use crate::plugin::path_rewrite::PathRewritePlugin;
 use crate::prelude::*;
